Modify clickSafe default.
From https://github.com/FormidableLabs/nuka-carousel/pull/190

When both swiping and dragging are disabled, clickSafe isn't needed
yet it is set to true by default because the handleSwipe function
is never called to unset it after a click. This means the click event is
always blocked.

This change replaces that behaviour by defaulting clickSafe to false and
enabling it if either dragging or swiping.